✨ Implement excluding detections from known contracts
