+++ wake/cli/detect.py
@@ -387,137 +387,6 @@ class DetectCli(click.RichGroup):  # pyright: ignore reportPrivateImportUsage
         super().invoke(ctx)
 
 
-class Location(NamedTuple):
-    source_unit_name: str
-    start: int
-    end: int
-
-    def overlaps(self, other: "Location") -> bool:
-        return (
-            self.source_unit_name == other.source_unit_name
-            and self.start <= other.end
-            and self.end >= other.start
-        )
-
-
-def find_relevant_locations(build: ProjectBuild, contract_name: str) -> Set[Location]:
-    from itertools import chain
-
-    from wake.ir import ContractDefinition, ExternalReference, IdentifierPathPart, IrAbc
-
-    locations = set()
-    # collect all relevant contracts, then find structs, events, etc. referenced by them
-    contracts = set()
-
-    for source_unit in build.source_units.values():
-        for contract in source_unit.contracts:
-            if contract.name != contract_name:
-                continue
-
-            # always include global using for directives
-            for using_for in source_unit.using_for_directives:
-                if using_for.library_name is not None:
-                    contracts.add(using_for.library_name.referenced_declaration)
-                    locations.add(
-                        Location(
-                            using_for.library_name.referenced_declaration.source_unit.source_unit_name,
-                            *using_for.library_name.referenced_declaration.byte_location,
-                        )
-                    )
-
-                locations.update(
-                    [
-                        Location(
-                            f.referenced_declaration.source_unit.source_unit_name,
-                            *f.referenced_declaration.byte_location,
-                        )
-                        for f in using_for.functions
-                    ]
-                )
-
-            contracts.update(contract.linearized_base_contracts)
-            locations.update(
-                [
-                    Location(b.source_unit.source_unit_name, *b.byte_location)
-                    for b in contract.linearized_base_contracts
-                ]
-            )
-
-            for using_for in contract.using_for_directives:
-                if using_for.library_name is not None:
-                    contracts.add(using_for.library_name.referenced_declaration)
-                    locations.add(
-                        Location(
-                            using_for.library_name.referenced_declaration.source_unit.source_unit_name,
-                            *using_for.library_name.referenced_declaration.byte_location,
-                        )
-                    )
-
-                locations.update(
-                    [
-                        Location(
-                            f.referenced_declaration.source_unit.source_unit_name,
-                            *f.referenced_declaration.byte_location,
-                        )
-                        for f in using_for.functions
-                    ]
-                )
-
-    def referenced_by(ref: Union[IrAbc, ExternalReference, IdentifierPathPart]) -> bool:
-        if isinstance(ref, IdentifierPathPart):
-            ref = ref.underlying_node
-        elif isinstance(ref, ExternalReference):
-            ref = ref.inline_assembly
-
-        while not isinstance(ref, ContractDefinition):
-            if ref.parent is None:
-                return False
-            ref = ref.parent
-
-        return ref in contracts
-
-    for source_unit in build.source_units.values():
-        for decl in chain(
-            source_unit.declared_variables,
-            source_unit.enums,
-            source_unit.functions,
-            source_unit.structs,
-            source_unit.errors,
-            source_unit.user_defined_value_types,
-            source_unit.events,
-        ):
-            if any(referenced_by(r) for r in decl.references):
-                locations.add(
-                    Location(source_unit.source_unit_name, *decl.byte_location)
-                )
-
-    return locations
-
-
-def is_detection_relevant(
-    detection: DetectorResult, relevant_locations: Set[Location]
-) -> bool:
-    detection_locations = set()
-
-    def add_detection_location(detection: Detection):
-        detection_locations.add(
-            Location(
-                detection.ir_node.source_unit.source_unit_name,
-                *detection.ir_node.byte_location,
-            )
-        )
-        if detection.subdetections_mandatory:
-            for subdetection in detection.subdetections:
-                add_detection_location(subdetection)
-
-    add_detection_location(detection.detection)
-
-    return any(
-        any(detection_location.overlaps(l) for l in relevant_locations)
-        for detection_location in detection_locations
-    )
-
-
 async def detect_(
     config: WakeConfig,
     no_artifacts: bool,
@@ -526,6 +395,7 @@ async def detect_(
     theme: str,
     watch: bool,
     ignore_disable_overrides: bool,
+    exclude_known_contracts: bool,
     import_json: Optional[str],
 ):
     import glob
@@ -543,6 +413,8 @@ async def detect_(
         DetectorConfidence,
         DetectorImpact,
         detect,
+        find_relevant_locations,
+        is_detection_relevant,
         print_detection,
     )
     from wake.detectors.utils import create_sarif_log
@@ -650,6 +522,7 @@ async def detect_(
             default_min_impact=default_min_impact,  # pyright: ignore reportGeneralTypeIssues
             default_min_confidence=default_min_confidence,  # pyright: ignore reportGeneralTypeIssues
             extra=extra,
+            strip_in_known_contracts=exclude_known_contracts,
         )
 
         extra.clear()
@@ -995,6 +868,12 @@ async def detect_(
     help="Print detections even if disabled with // wake-disable-* comments.",
 )
 @click.option(
+    "--exclude-known-contracts/--no-exclude-known-contracts",
+    is_flag=True,
+    default=True,
+    help="Exclude detections in known contracts.",
+)
+@click.option(
     "--ignore-path",
     "ignore_paths",
     multiple=True,
@@ -1113,6 +992,7 @@ def run_detect(
     theme: str,
     watch: bool,
     ignore_disable_overrides: bool,
+    exclude_known_contracts: bool,
     ignore_paths: Tuple[str],
     detect_exclude_paths: Tuple[str],
     exclude: Tuple[str],
@@ -1198,6 +1078,7 @@ def run_detect(
             theme,
             watch,
             ignore_disable_overrides,
+            exclude_known_contracts,
             import_json,
         )
     )
